webots_stuff/webots-scripts/llm-controlled-script/requirements.txt
opencv-python
numpy
python-dotenv
pymavlink
openai
google-genai